# api-contract/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Learning Hub API
  version: 1.0.0
  description: Product-scoped API contract for learning resources and tags.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.learninghub.local
    description: Contract reference server
security: []
paths:
  /api/v1/{product}/resources:
    get:
      operationId: listResources
      tags: [Resources]
      summary: List resources
      parameters:
        - $ref: "#/components/parameters/ProductPathParam"
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [video, pdf, article, all]
        - name: tags
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated tags
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Offset encoded as string
        - name: limit
          in: query
          required: false
          schema:
            type: string
          description: Items per page
      responses:
        "200":
          description: Resources list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResourceResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

    post:
      operationId: createResource
      tags: [Resources]
      summary: Create resource
      parameters:
        - $ref: "#/components/parameters/ProductPathParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateResourceRequest"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

  /api/v1/{product}/resources/{id}:
    get:
      operationId: getResourceById
      tags: [Resources]
      summary: Get resource by ID
      parameters:
        - $ref: "#/components/parameters/ProductPathParam"
        - $ref: "#/components/parameters/ResourceIdPathParam"
      responses:
        "200":
          description: Resource details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

    patch:
      operationId: updateResource
      tags: [Resources]
      summary: Update resource
      parameters:
        - $ref: "#/components/parameters/ProductPathParam"
        - $ref: "#/components/parameters/ResourceIdPathParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateResourceRequest"
      responses:
        "200":
          description: Resource updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

    delete:
      operationId: deleteResource
      tags: [Resources]
      summary: Delete resource
      parameters:
        - $ref: "#/components/parameters/ProductPathParam"
        - $ref: "#/components/parameters/ResourceIdPathParam"
      responses:
        "200":
          description: Resource deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResourceResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

  /api/v1/{product}/tags:
    get:
      operationId: listTags
      tags: [Tags]
      summary: Get tags
      parameters:
        - $ref: "#/components/parameters/ProductPathParam"
      responses:
        "200":
          description: Tags list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

components:
  parameters:
    ProductPathParam:
      name: product
      in: path
      required: true
      schema:
        type: string
      description: Product scope identifier
    ResourceIdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier

  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ResourceType:
      type: string
      enum: [video, pdf, article]

    Resource:
      type: object
      required:
        - id
        - title
        - description
        - type
        - url
        - tags
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/ResourceType"
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Tag:
      type: object
      required: [name, usageCount]
      properties:
        name:
          type: string
        usageCount:
          type: integer

    PaginatedResourceResponse:
      type: object
      required: [data, hasMore]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        hasMore:
          type: boolean
        nextCursor:
          type: string

    CreateResourceRequest:
      type: object
      required: [title, description, type]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/ResourceType"
        tags:
          type: string
          description: Comma-separated tags
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        file:
          type: string
          format: binary
        thumbnail:
          type: string
          format: binary

    UpdateResourceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/ResourceType"
        tags:
          type: string
          description: Comma-separated tags
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        file:
          type: string
          format: binary
        thumbnail:
          type: string
          format: binary

    DeleteResourceResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: string