# backend/contract/types.gen.go
// Package contract provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package contract

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ResourceType.
const (
	ResourceTypeArticle ResourceType = "article"
	ResourceTypePdf     ResourceType = "pdf"
	ResourceTypeVideo   ResourceType = "video"
)

// Valid indicates whether the value is a known member of the ResourceType enum.
func (e ResourceType) Valid() bool {
	switch e {
	case ResourceTypeArticle:
		return true
	case ResourceTypePdf:
		return true
	case ResourceTypeVideo:
		return true
	default:
		return false
	}
}

// Defines values for ListResourcesParamsType.
const (
	ListResourcesParamsTypeAll     ListResourcesParamsType = "all"
	ListResourcesParamsTypeArticle ListResourcesParamsType = "article"
	ListResourcesParamsTypePdf     ListResourcesParamsType = "pdf"
	ListResourcesParamsTypeVideo   ListResourcesParamsType = "video"
)

// Valid indicates whether the value is a known member of the ListResourcesParamsType enum.
func (e ListResourcesParamsType) Valid() bool {
	switch e {
	case ListResourcesParamsTypeAll:
		return true
	case ListResourcesParamsTypeArticle:
		return true
	case ListResourcesParamsTypePdf:
		return true
	case ListResourcesParamsTypeVideo:
		return true
	default:
		return false
	}
}

// CreateResourceRequest defines model for CreateResourceRequest.
type CreateResourceRequest struct {
	Description string              `json:"description"`
	File        *openapi_types.File `json:"file,omitempty"`

	// Tags Comma-separated tags
	Tags         *string             `json:"tags,omitempty"`
	Thumbnail    *openapi_types.File `json:"thumbnail,omitempty"`
	ThumbnailUrl *string             `json:"thumbnailUrl,omitempty"`
	Title        string              `json:"title"`
	Type         ResourceType        `json:"type"`
	Url          *string             `json:"url,omitempty"`
}

// DeleteResourceResponse defines model for DeleteResourceResponse.
type DeleteResourceResponse struct {
	Message string `json:"message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Details *string `json:"details,omitempty"`
	Error   string  `json:"error"`
	Message *string `json:"message,omitempty"`
}

// PaginatedResourceResponse defines model for PaginatedResourceResponse.
type PaginatedResourceResponse struct {
	Data       []Resource `json:"data"`
	HasMore    bool       `json:"hasMore"`
	NextCursor *string    `json:"nextCursor,omitempty"`
}

// Resource defines model for Resource.
type Resource struct {
	CreatedAt    time.Time    `json:"createdAt"`
	Description  string       `json:"description"`
	Id           string       `json:"id"`
	Tags         []string     `json:"tags"`
	ThumbnailUrl *string      `json:"thumbnailUrl,omitempty"`
	Title        string       `json:"title"`
	Type         ResourceType `json:"type"`
	UpdatedAt    time.Time    `json:"updatedAt"`
	Url          string       `json:"url"`
}

// ResourceType defines model for ResourceType.
type ResourceType string

// Tag defines model for Tag.
type Tag struct {
	Name       string `json:"name"`
	UsageCount int    `json:"usageCount"`
}

// UpdateResourceRequest defines model for UpdateResourceRequest.
type UpdateResourceRequest struct {
	Description *string             `json:"description,omitempty"`
	File        *openapi_types.File `json:"file,omitempty"`

	// Tags Comma-separated tags
	Tags         *string             `json:"tags,omitempty"`
	Thumbnail    *openapi_types.File `json:"thumbnail,omitempty"`
	ThumbnailUrl *string             `json:"thumbnailUrl,omitempty"`
	Title        *string             `json:"title,omitempty"`
	Type         *ResourceType       `json:"type,omitempty"`
	Url          *string             `json:"url,omitempty"`
}

// ProductPathParam defines model for ProductPathParam.
type ProductPathParam = string

// ResourceIdPathParam defines model for ResourceIdPathParam.
type ResourceIdPathParam = string

// ListResourcesParams defines parameters for ListResources.
type ListResourcesParams struct {
	Type *ListResourcesParamsType `form:"type,omitempty" json:"type,omitempty"`

	// Tags Comma-separated tags
	Tags   *string `form:"tags,omitempty" json:"tags,omitempty"`
	Search *string `form:"search,omitempty" json:"search,omitempty"`

	// Cursor Offset encoded as string
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Items per page
	Limit *string `form:"limit,omitempty" json:"limit,omitempty"`
}

// ListResourcesParamsType defines parameters for ListResources.
type ListResourcesParamsType string

// CreateResourceMultipartRequestBody defines body for CreateResource for multipart/form-data ContentType.
type CreateResourceMultipartRequestBody = CreateResourceRequest

// UpdateResourceMultipartRequestBody defines body for UpdateResource for multipart/form-data ContentType.
type UpdateResourceMultipartRequestBody = UpdateResourceRequest